src/backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pymysql
bcrypt
PyJWT
email-validator
fastapi-mail
python-dotenv
jinja2
